--- src/main/java/org/example/logistics/CategoriesVO.java
@@ -1,11 +1,0 @@
-package org.example.logistics;
-
-import lombok.Data;
-
-@Data
-public class CategoriesVO {
-    private int categoryId;
-    private String name;
-    private String description;
-}
-
