--- src/main/java/org/example/logistics/ManufacturersVO.java
@@ -1,12 +1,0 @@
-package org.example.logistics;
-
-import lombok.Data;
-
-@Data
-
-public class ManufacturersVO {
-    private int manufacturerId;
-    private String name;
-    private String location;
-    private String contact;
-}
