--- src/main/java/org/example/logistics/WarehousesVO.java
@@ -1,11 +1,0 @@
-package org.example.logistics;
-
-import lombok.Data;
-
-@Data
-public class WarehousesVO {
-    private int warehouseId;
-    private String name;
-    private String location;
-    private int capacity; // 창고 용량
-}
