+++ src/main/java/org/example/logistics/administrators/AdministratorsVO.java
@@ -1,0 +1,11 @@
+package org.example.logistics.administrators;
+
+import lombok.Data;
+
+@Data
+public class AdministratorsVO {
+    private int admin_id;
+    private String user_id;
+    private String password;
+    private String role;
+}
