+++ src/main/java/org/example/logistics/branches/BranchesOutgoingOrdersProductsVO.java
@@ -1,9 +1,9 @@
-package org.example.logistics;
+package org.example.logistics.branches;
 
 import lombok.Data;
 
 @Data
-public class BranchesVO {
+public class BranchesOutgoingOrdersVO {
     private int branchId;
     private String name;
     private String location;
